+++ src/App.vue
@@ -10,9 +10,14 @@
           width="32"
         />
 
-        <v-toolbar-title @click="$router.push('/')" class="clickable">
-          <span class="d-none d-sm-inline">Gunchained Arena</span>
-          <span class="d-inline d-sm-none">GU</span>
+        <v-toolbar-title>
+          <span @click="$router.push('/')" class="clickable">
+            <span class="d-none d-sm-inline">Gunchained</span>
+            <span class="d-inline d-sm-none">GU</span>
+          </span>
+          <span class="d-none d-sm-inline">
+             {{ title }}
+          </span>
         </v-toolbar-title>
 
         <v-chip
@@ -31,17 +36,10 @@
           </span>
         </v-chip>
       </div>
+
       <v-spacer></v-spacer>
 
       <v-toolbar-items class="d-flex align-center">
-        <div v-if="!user">
-          <Tooltip tooltip="Sign in" bottom>
-            <v-btn to="/login" text fab small class="text--secondary">
-              <v-icon color="white">mdi-login</v-icon>
-            </v-btn>
-          </Tooltip>
-        </div>
-
         <div v-if="user" class="d-flex align-center ">
           <v-switch
             dark
@@ -62,15 +60,35 @@
             </v-avatar>
             Profile
           </v-chip>
-
-          <Tooltip tooltip="Log out" bottom>
-            <v-btn @click="logout" text fab small class="text--secondary">
-              <v-icon color="white">mdi-logout</v-icon>
-            </v-btn>
-          </Tooltip>
         </div>
       </v-toolbar-items>
+
+      <v-app-bar-nav-icon @click.stop="drawer = !drawer" class="mr-1"></v-app-bar-nav-icon>
     </v-app-bar>
+
+    <v-navigation-drawer v-model="drawer" app dark temporary right width="200">
+      <v-list dense nav>
+        <v-list-item link v-for="link in navigation" :key="link.label" :to="link.route">
+          <v-list-item-icon>
+            <v-icon>{{ link.icon }}</v-icon>
+          </v-list-item-icon>
+
+          <v-list-item-content>
+            <v-list-item-title>{{ link.label }}</v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+
+        <v-list-item link @click="logout" v-if="user">
+          <v-list-item-icon>
+            <v-icon>mdi-logout</v-icon>
+          </v-list-item-icon>
+
+          <v-list-item-content>
+            <v-list-item-title>Log out</v-list-item-title>
+          </v-list-item-content>
+        </v-list-item>
+      </v-list>
+    </v-navigation-drawer>
 
     <v-main>
       <v-alert
@@ -129,7 +147,7 @@ export default {
     FooterBtn
   },
   computed: {
-    ...mapState(['user', 'app', 'status']),
+    ...mapState(['user', 'app', 'status', 'title']),
     ...mapGetters(['appStatus', 'isAvailableEnabled']),
     ...mapGetters({
       player: 'playerObject'
@@ -139,6 +157,11 @@ export default {
     },
     pluralizedUserString() {
       return this.activeUserCount === 1 ? 'user' : 'users'
+    },
+    navigation() {
+      const links = this.links
+      if (!this.user) links.pop
+      return links
     }
   },
   created() {
@@ -153,11 +176,30 @@ export default {
   },
   data() {
     return {
-      version: '0.1.4'
+      version: '0.1.5',
+      drawer: false,
+      links: [
+        {
+          label: "Arena",
+          icon: "mdi-fencing",
+          route: "/"
+        },
+        {
+          label: "Videos",
+          icon: 'mdi-video',
+          route: "/videos"
+        },
+        {
+          label: "Sign in",
+          icon: "mdi-login",
+          route: "/login"
+        }
+      ]
     }
   },
   methods: {
     logout() {
+      console.log("Logging out")
       firebase
         .auth()
         .signOut()