+++ src/views/Home.vue
@@ -8,6 +8,9 @@ import { mapState } from 'vuex'
 
 export default {
   name: 'Home',
+  mounted() {
+    this.$store.dispatch('setTitle', 'Arena')
+  },
   components: {
     ChallengeList
   },